One more file uploaded
Hi again and Happy new week! One more file uploaded here, part of source file in English already sent to you in the beginning, translated to Spanish. The entire book will be translated for YOU AND LOCALE, to help you both, and I'll always try to work for you, and herein Locale also in this case, my pleasure, as in the past, 
Kindly, next time with more content, 
Kat